requirements.txt
google-adk
google-genai
httpx
feedparser
python-dotenv